refactor: s/bn_helpers/secure/g, move definitions to TU, header internal
